v0.1
A very basic, rough and ready piece of Javascript and HTML code that can be customised and projected in an exam hall to display the current time, with entry boxes for the exam start time, a percentage of extra time for pupils with SEN requirements, and have the standard end time and extra time end time calculated and displayed automatically.

It's not flash, but it works. Current time is taken from your system time so please make sure that's correct before using.

Released under the FreeBSD Licence.